refactor: file organization and dedicated files for functions
