+++ .gitignore
@@ -1,3 +1,3 @@
 .venv
 .vscode
-__pycache__
+*/__pycache__
